[web-dashboard] delete visible only for admin

+++ web/src/pages/dashboard/Employees.tsx
@@ -1,13 +1,16 @@
 import * as Chakra from "@chakra-ui/react";
 import { EmployeeTable } from "@features";
-import { useEmployeeStore } from "@stores";
+import { useEmployeeStore, useUserInfoStore } from "@stores";
+import { UserRole } from "@http";
 import { fetchChain } from "@fetchChain";
 
 const EmployeesPage = () => {
 	const { employees } = useEmployeeStore();
 	const fetch = new fetchChain();
+	const { role } = useUserInfoStore();
 
 	const handleDelete = () => {
+		if(role !== UserRole.ADMIN) return;
 		const ids = employees.filter((item) => item.checked).map((item) => item.id);
 		for(let v of ids) {
 			fetch.deleteEmployee(v).then();
@@ -17,9 +20,9 @@ const EmployeesPage = () => {
     return (
         <Chakra.Box p={6}>
             <Chakra.Heading mb={4}>Employees</Chakra.Heading>
-			<Chakra.Button bg="red.500" color="white" _hover={{ bg: "#0A76E6" }} fontWeight="semibold" onClick={handleDelete}>
+			{role === UserRole.ADMIN ? <Chakra.Button bg="red.500" color="white" _hover={{ bg: "#0A76E6" }} fontWeight="semibold" onClick={handleDelete}>
 				Delete
-			</Chakra.Button>
+			</Chakra.Button> : <></>}
 			<Chakra.Box w="50%"><EmployeeTable/></Chakra.Box>
         </Chakra.Box>
     );
